fix: 수정사항 반영

+++ src/main/java/com/sprint/mission/findex/indexdata/dto/response/CursorPageResponseIndexDataDto.java
@@ -9,8 +9,8 @@ public record CursorPageResponseIndexDataDto<T> (
         List<T> content,
         String nextCursor,
         Long nextIdAfter,
-        Integer size,
-        Long totalElements,
+        int size,
+        int totalElements,
         Boolean hasNext
 ) {
 }
